requirements.txt
pyperclip==1.7.0
pycryptodome==3.6.6